pyproject.toml
[project]
name = "private-pdf-rag"
version = "0.1.0"
description = "RAG application for chatting with PDF documents"
authors = [
    {name = "Mathanraj-Sharma", email = "rvmmathanraj@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.28.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.4.15",
    "ollama>=0.1.7",
    "pypdf2>=3.0.1",
    "sentence-transformers>=2.2.2",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.7.4",
    "tiktoken>=0.5.1",
    "streamlit-chat>=0.1.1",
    "python-multipart>=0.0.6",
    "debugpy>=1.8.14",
    "snakeviz>=2.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
